+++ internal/runners/deploy/deploy_test.go
@@ -4,12 +4,10 @@ import (
 	"os"
 	"path/filepath"
 	"reflect"
-	"sort"
 	"testing"
 
 	"github.com/ActiveState/cli/internal/failures"
 	"github.com/ActiveState/cli/internal/output"
-	"github.com/ActiveState/cli/internal/runners/activate"
 	"github.com/ActiveState/cli/internal/subshell"
 	"github.com/ActiveState/cli/internal/testhelpers/outputhelper"
 	"github.com/ActiveState/cli/pkg/platform/runtime"
@@ -225,59 +223,6 @@ func Test_report(t *testing.T) {
 
 			if !reflect.DeepEqual(report.BinaryDirectories, tt.wantBinary) {
 				t.Errorf("Expected bins to be the same. Want: %v, got: %v", tt.wantBinary, report.BinaryDirectories)
-			}
-		})
-	}
-}
-
-func Test_uniqueExes(t *testing.T) {
-	tests := []struct {
-		name    string
-		bins    []string
-		pathext string
-		want    []string
-	}{
-		{
-			"Returns same bins",
-			[]string{"path1/a", "path2/b", "path3/c"},
-			"",
-			[]string{"path1/a", "path2/b", "path3/c"},
-		},
-		{
-			"Returns exe prioritized",
-			[]string{"path1/a.exe", "path2/a.cmd", "path3/c"},
-			".exe;.cmd",
-			[]string{"path1/a.exe", "path3/c"},
-		},
-		{
-			"Returns cmd prioritized by PATH",
-			[]string{"path1/a.exe", "path2/a.cmd", "path2/c"},
-			".cmd;.exe",
-			[]string{"path1/a.exe", "path2/c"},
-		},
-		{
-			"Returns cmd prioritized by PATHEXT",
-			[]string{"path1/a.exe", "path1/a.cmd", "path2/c"},
-			".cmd;.exe",
-			[]string{"path1/a.cmd", "path2/c"},
-		},
-		{
-			"PATHEXT can be empty",
-			[]string{"path1/a", "path2/b", "path3/c"},
-			"",
-			[]string{"path1/a", "path2/b", "path3/c"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := activate.UniqueExes(tt.bins, tt.pathext)
-			if err != nil {
-				t.Errorf("uniqueExes error: %v", err)
-				t.FailNow()
-			}
-			sort.Strings(got)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("uniqueExes() = %v, want %v", got, tt.want)
 			}
 		})
 	}
